feat(reference): introduce resolve strategies framework
Refs https://github.com/swagger-api/oss-planning/issues/133
