+++ apidom/packages/apidom-reference/src/options/index.ts
@@ -1,15 +1,16 @@
-import { mergeDeepRight } from 'ramda';
+import FileResolver from '../resolvers/FileResolver';
+import HttpResolverAxios from '../resolvers/HttpResolverAxios';
 
-import FileResolver from './resolvers/FileResolver';
-import HttpResolverAxios from './resolvers/HttpResolverAxios';
-import OpenApiJson3_1Parser from './parsers/apidom-reference-parser-openapi-json-3-1';
-import OpenApiYaml3_1Parser from './parsers/apidom-reference-parser-openapi-yaml-3-1';
-import AsyncApiJson2_0Parser from './parsers/apidom-reference-parser-asyncapi-json-2-0';
-import AsyncApiYaml2_0Parser from './parsers/apidom-reference-parser-asyncapi-yaml-2-0';
-import JsonParser from './parsers/apidom-reference-parser-json';
-import YamlParser from './parsers/apidom-reference-parser-yaml';
+import OpenApi3_1ResolveStrategy from '../resolve-strategies/openapi-3-1';
 
-import { ReferenceOptions as IReferenceOptions } from './types';
+import OpenApiJson3_1Parser from '../parsers/apidom-reference-parser-openapi-json-3-1';
+import OpenApiYaml3_1Parser from '../parsers/apidom-reference-parser-openapi-yaml-3-1';
+import AsyncApiJson2_0Parser from '../parsers/apidom-reference-parser-asyncapi-json-2-0';
+import AsyncApiYaml2_0Parser from '../parsers/apidom-reference-parser-asyncapi-yaml-2-0';
+import JsonParser from '../parsers/apidom-reference-parser-json';
+import YamlParser from '../parsers/apidom-reference-parser-yaml';
+
+import { ReferenceOptions as IReferenceOptions } from '../types';
 
 const defaultOptions: IReferenceOptions = {
   parse: {
@@ -50,24 +51,25 @@ const defaultOptions: IReferenceOptions = {
       HttpResolverAxios({ timeout: 5000, redirects: 5, withCredentials: false }),
     ],
     /**
+     * If set to a specific resolve strategy, loop for suitable resolve strategy will be skipped
+     * and this strategy will be used directly.
+     */
+    strategy: null,
+    /**
+     * Determines strategies how References are identified and processed by resolvers.
+     * Strategy is determined by media type.
+     *
+     * You can add additional resolver strategies of your own, replace an existing one with
+     * your own implementation, or remove any resolver by removing it from the list.
+     */
+    strategies: [OpenApi3_1ResolveStrategy()],
+    /**
      * Determines whether external references will be resolved.
      * If this option is disabled, then none of above resolvers will be called.
      * Instead, external references will simply be ignored.
-     *
-     * @type {boolean}
      */
     external: true,
   },
 };
-
-/**
- * Algorithm for deep merging options.
- */
-export const merge = mergeDeepRight;
-
-/**
- * Algorithm for deep merging options with default ones.
- */
-export const mergeWithDefaults = mergeDeepRight(defaultOptions);
 
 export default defaultOptions;
