+++ apidom/packages/apidom-reference/src/options/util.ts
@@ -1,0 +1,7 @@
+import { mergeDeepRight } from 'ramda';
+
+/**
+ * Algorithm for deep merging options.
+ */
+// eslint-disable-next-line import/prefer-default-export
+export const merge = mergeDeepRight;
