+++ apidom/packages/apidom-reference/src/resolve-strategies/ResolveStrategy.ts
@@ -1,0 +1,17 @@
+import stampit from 'stampit';
+
+import { ResolveStrategy as IResolveStrategy } from '../types';
+import { NotImplementedError } from '../util/errors';
+
+const ResolveStrategy: stampit.Stamp<IResolveStrategy> = stampit({
+  methods: {
+    canResolve() {
+      return false;
+    },
+    async resolve(): Promise<never> {
+      throw new NotImplementedError();
+    },
+  },
+});
+
+export default ResolveStrategy;
