+++ apidom/packages/apidom-reference/src/resolve-strategies/openapi-3-1/reference-objects/predicates.ts
