--- apidom/packages/apidom-reference/src/strategies/openapi-3-1/resolve/index.ts
@@ -1,1 +1,0 @@
-export {};
