+++ apidom/packages/apidom-reference/test/resolve-strategies/openapi-3-1/reference-objects/fixtures/external-reference-depth-4/ex1.json
