+++ apidom/packages/apidom-reference/test/resolve-strategies/openapi-3-1/reference-objects/fixtures/nested-external-reference-depth-4/nesting1/ex1.json
