+++ apidom/packages/apidom-reference/test/resolve-strategies/openapi-3-1/reference-objects/fixtures/nested-external-reference-depth-4/nesting1/nesting2/ex2.json
