+++ apidom/packages/apidom-reference/test/resolve-strategies/openapi-3-1/reference-objects/fixtures/no-external-reference/root.json
